\name{wxfR-package}
\alias{wxfR-package}
\alias{wxfR}
\docType{package}
\title{
\packageTitle{wxfR}
}
\description{
\packageDescription{wxfR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wxfR}
\packageIndices{wxfR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{wxfR}

Maintainer: \packageMaintainer{wxfR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
